实现websocket和ssl，测试了websocket

+++ sylar/stream.cc
@@ -6,10 +6,10 @@ namespace sylar
 int Stream::readFixSize(void *buffer, size_t length)
 {
     size_t offset = 0;
-    size_t left = length;
+    int64_t left = length;
     while (left > 0)
     {
-        size_t len = read((char *)buffer + offset, left);
+        int64_t len = read((char *)buffer + offset, left);
         if (len <= 0)
         {
             return len;
@@ -22,10 +22,10 @@ int Stream::readFixSize(void *buffer, size_t length)
 
 int Stream::readFixSize(ByteArray::ptr ba, size_t length)
 {
-    size_t left = length;
+    int64_t left = length;
     while (left > 0)
     {
-        size_t len = read(ba, left);
+        int64_t len = read(ba, left);
         if (len <= 0)
         {
             return len;
@@ -38,10 +38,10 @@ int Stream::readFixSize(ByteArray::ptr ba, size_t length)
 int Stream::writeFixSize(const void *buffer, size_t length)
 {
     size_t offset = 0;
-    size_t left = length;
+    int64_t left = length;
     while (left > 0)
     {
-        size_t len = write((const char *)buffer + offset, left);
+        int64_t len = write((const char *)buffer + offset, left);
         if (len <= 0)
         {
             return len;
@@ -54,10 +54,10 @@ int Stream::writeFixSize(const void *buffer, size_t length)
 
 int Stream::writeFixSize(ByteArray::ptr ba, size_t length)
 {
-    size_t left = length;
+    int64_t left = length;
     while (left > 0)
     {
-        size_t len = write(ba, left);
+        int64_t len = write(ba, left);
         if (len <= 0)
         {
             return len;
